tested both netcdf and ascii input files
