+++ examples/storm-surge/isaac_ascii_data/isaac.info
@@ -1,0 +1,8 @@
+## Control file for OWI ascii wind and pressure data files ##
+## Edit Files to include all wind and pressure data per storm ##
+## Ensure data remains in the below format ##
+
+201208290000                          =: landfall_date    # Date of landfall for the storm
+"/home/catherinej/clawpack/geoclaw/examples/storm-surge/isaac_ascii_data/isaac.WIN"                           =: wind_field       # path to wind forcing file
+'/home/catherinej/clawpack/geoclaw/examples/storm-surge/isaac_ascii_data/isaac.PRE'                           =: pressure_field   # path to pressure forcing file
+0                                     =: regional_forcing # 0 for only basin forcing, 1 for regional
