+++ examples/storm-surge/isaac_netcdf_data/.gitignore
@@ -1,0 +1,1 @@
+isaac.storm
