--- examples/storm-surge/isaac_syn/.gitignore
@@ -1,1 +1,0 @@
-isaac.storm
